add cart/fav buttons component
