+++ src/components/SliderCart/SliderCart.module.scss
@@ -2,6 +2,8 @@
 
 .cart{
   width: min-content;
+  background-color: white;
+  border-radius: 5px;
 }
 .image{
   width: 280px;
